Minor HomeHero > HomeIntro component and variable renaming
