+++ app/layout/home/intro/intro-animation/index.ts
